+++ src/Oxrun/Application/oxrun-light.php
@@ -12,4 +12,5 @@ declare(strict_types=1);
 $application = new \Symfony\Component\Console\Application('oxrun-light', '0.2');
 $application->add(new \Oxrun\Command\Misc\RegisterCommand());
 $application->add(new \Oxrun\Command\Misc\PhpstormMetadataCommand());
+$application->add(new \Oxrun\Command\Misc\GenerateDocumentationCommand());
 $application->run();