+++ src/Oxrun/Command/Misc/GenerateDocumentationCommand.php
@@ -2,10 +2,17 @@
 
 namespace Oxrun\Command\Misc;
 
+use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
+use OxidEsales\EshopCommunity\Internal\Framework\Console\ExecutorInterface;
+use OxidEsales\Facts\Facts;
+use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\NullOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Tester\CommandTester;
+use Webmozart\PathUtil\Path;
 
 /**
  * Class GenerateDocumentationCommand
@@ -27,6 +34,8 @@ class GenerateDocumentationCommand extends Command
         '* Is required: no',
         '* Accept value: yes',
     );
+
+    protected $isOeConsoleExecuted = false;
 
     /**
      * Configures the current command.
@@ -53,11 +62,10 @@ class GenerateDocumentationCommand extends Command
             $this->skipLines
         );
 
-        $availableCommands = array_keys($this->getApplication()->all());
+        $availableCommands = array_keys($this->getOeConsoleApp()->all());
 
-
-        $availableCommands = array_filter($availableCommands, function ($commandName) {
-            if (in_array($commandName, $this->skipCommands) || preg_match('/^oe:/', $commandName) !== false) {
+        $availableCommands = array_filter($availableCommands, function ($commandName) use ($output) {
+            if (in_array($commandName, $this->skipCommands) || preg_match('/^oe:/', $commandName) === 1) {
                 return false;
             }
             return true;
@@ -65,7 +73,7 @@ class GenerateDocumentationCommand extends Command
 
         sort($availableCommands);
 
-        $this->writeToc($output);
+        $this->writeToc($output, $availableCommands);
 
         $output->writeLn(PHP_EOL);
 
@@ -74,11 +82,30 @@ class GenerateDocumentationCommand extends Command
     }
 
     /**
+     * @return \Symfony\Component\Console\Application
+     */
+    public function getOeConsoleApp(): Application
+    {
+        $container = ContainerFactory::getInstance()->getContainer();
+        /** @var Application $app */
+        $app = $container->get('oxid_esales.console.symfony.component.console.application');
+
+        if ($this->isOeConsoleExecuted === false) {
+            include_once Path::join((new Facts())->getSourcePath(), 'bootstrap.php');
+            $app->setAutoExit(false);
+            $container->get(ExecutorInterface::class)->execute(new ArrayInput([]), new NullOutput());
+            $this->isOeConsoleExecuted = true;
+        }
+
+        return $app;
+    }
+
+    /**
      * @param OutputInterface $output
      */
-    protected function writeToc(OutputInterface $output)
+    protected function writeToc(OutputInterface $output, array $availableCommands)
     {
-        $command = $this->getApplication()->find('list');
+        $command = $this->getOeConsoleApp()->find('list');
         $commandTester = new CommandTester($command);
         $commandTester->execute(
             array(
@@ -86,26 +113,42 @@ class GenerateDocumentationCommand extends Command
                 '--format' => 'json'
             )
         );
+
         $commandOutput = $commandTester->getDisplay();
         $commandOutput = json_decode($commandOutput);
         $output->writeLn("Available commands");
         $output->writeLn("==================" . PHP_EOL);
 
         $description = [];
-        array_walk($commandOutput->commands, function ($item) use (&$description) {
-            $description[$item->name] = $item->description;
-        });
+        array_walk(
+            $commandOutput->commands,
+            function ($item) use (&$description, $availableCommands) {
+                if (in_array($item->name, $availableCommands)) {
+                    $description[$item->name] = $item->description;
+                }
+            }
+        );
 
         foreach ($commandOutput->namespaces as $namespace) {
+            $title = "";
             if ($namespace->id != '_global') {
-                $output->writeLn("##### $namespace->id");
+                $title = $namespace->id;
             };
+            $links = [];
             foreach ($namespace->commands as $command) {
-                if (in_array($command, $this->skipCommands)) {
+                if (in_array($command, $this->skipCommands) || !in_array($command, $availableCommands)) {
                     continue;
                 }
-                $link = sprintf('- [%s](#%s)   %s', $command, str_replace(':', '', $command), $description[$command]);
-                $output->writeLn("  $link");
+                $links[] = sprintf(
+                    '  - [%s](#%s)   %s',
+                    $command,
+                    str_replace(':', '', $command),
+                    $description[$command]
+                );
+            }
+            if (!empty($links)) {
+                $output->writeLn("##### $title");
+                array_map([$output, 'writeLn'], $links);
             }
         };
     }
@@ -116,8 +159,9 @@ class GenerateDocumentationCommand extends Command
      */
     protected function writeCommandUsage(OutputInterface $output, $availableCommands)
     {
-        $command = $this->getApplication()->find('help');
+        $command = $this->getOeConsoleApp()->find('help');
         $commandTester = new CommandTester($command);
+        $applicationOptions = array_keys($this->getOeConsoleApp()->getDefinition()->getOptions());
 
         foreach ($availableCommands as $commandName) {
             $commandTester->execute(
@@ -127,14 +171,21 @@ class GenerateDocumentationCommand extends Command
                     '--format' => 'md'
                 )
             );
+
             $commandOutput = $commandTester->getDisplay();
-            $commandOutput = substr($commandOutput, 0, strpos($commandOutput, '**help:**'));
+
+            //Remove unnecessary information
             $commandOutput = str_replace($this->skipLines, '', $commandOutput);
 
-            $currentCommand = $this->getApplication()->find($commandName);
+            //Remove Title Option if has no Options
+            $currentCommand = $this->getOeConsoleApp()->find($commandName);
             if (count($currentCommand->getDefinition()->getOptions()) < 8) {
-                $commandOutput = str_replace('### Options:', '', $commandOutput);
+                $commandOutput = str_replace('### Options', '', $commandOutput);
             }
+
+            //Remove standard application options
+            $pattern = "/#### `--(?:" . implode('|', $applicationOptions) . ")[^#]+/";
+            $commandOutput = preg_replace($pattern, '', $commandOutput);
 
             $commandOutput = trim($commandOutput);
             $output->writeLn($commandOutput . PHP_EOL);